+++ react/Reviews.js
@@ -3,7 +3,7 @@ import React, { useContext, useEffect, useCallback, useReducer } from 'react'
 import { ProductContext } from 'vtex.product-context'
 import queryRatingSummary from './graphql/queries/queryRatingSummary.gql'
 import voteReviewQuery from './graphql/mutations/voteReview.gql'
-import getConfig from './graphql/getConfig.graphql'
+import getConfig from './graphql/getConfig.gql'
 import { withApollo, graphql } from 'react-apollo'
 
 import {
@@ -43,7 +43,7 @@ const options = [
 
 const filters = [
   {
-    label: 'Select a filter...',
+    label: 'All',
     value: '0',
   },
   {
@@ -99,7 +99,7 @@ const initialState = {
   count: 0,
   percentage: [],
   selected: 'Newest',
-  filter: 0,
+  filter: '0',
   paging: {},
   page: 0,
 }
@@ -307,7 +307,7 @@ const Reviews = props => {
   }
 
   return state.reviews.length ? (
-    <div className="review mw8 center ph5">
+    <div className="review mw8 center ph5" id="all-reviews">
       <h3 className="review__title t-heading-3 bb b--muted-5 mb5">Reviews</h3>
       <div className="review__rating">
         <div className="review__rating--stars dib relative v-mid mr2">
@@ -410,21 +410,21 @@ const Reviews = props => {
             Reviewed by {state.count}{' '}
             {state.count == 1 ? 'customer' : 'customers'}
           </h4>
-          <div className="mb7">
-            <Dropdown
-              options={options}
-              onChange={handleSort}
-              value={state.selected}
-              {...props}
-            />
-          </div>
-          <div className="mb7">
-            <Dropdown
-              options={filters}
-              onChange={handleFilter}
-              value={state.filter}
-              {...props}
-            />
+          <div className="flex mb7">
+            <div className="mr4">
+              <Dropdown
+                options={options}
+                onChange={handleSort}
+                value={state.selected}
+              />
+            </div>
+            <div className="">
+              <Dropdown
+                options={filters}
+                onChange={handleFilter}
+                value={state.filter}
+              />
+            </div>
           </div>
 
           <div className="mv5">
@@ -594,21 +594,21 @@ const Reviews = props => {
             Reviewed by {state.count}{' '}
             {state.count == 1 ? 'customer' : 'customers'}
           </h4>
-          <div className="mb7">
-            <Dropdown
-              options={options}
-              onChange={handleSort}
-              value={state.selected}
-              {...props}
-            />
-          </div>
-          <div className="mb7">
-            <Dropdown
-              options={filters}
-              onChange={handleFilter}
-              value={state.filter}
-              {...props}
-            />
+          <div className="flex mb7">
+            <div className="mr4">
+              <Dropdown
+                options={options}
+                onChange={handleSort}
+                value={state.selected}
+              />
+            </div>
+            <div>
+              <Dropdown
+                options={filters}
+                onChange={handleFilter}
+                value={state.filter}
+              />
+            </div>
           </div>
 
           <div className="mv5">
@@ -629,7 +629,7 @@ const Reviews = props => {
 
           <div className="review__comment bw2 bb b--muted-5 mb5 pb4">
             <h5 className="review__comment--user lh-copy mw9 t-heading-5 mv5">
-              No reviews found!
+              No reviews.
             </h5>
           </div>
         </div>